[tempSave] create 페이지 컴포넌트 분리 진행중 #49
